refactor: separate inventory management functions into .h and .c files
